dedicated barchart block implemented
